Rename grpc base client file name
