+++ services/assistant/app/adapters/output/grpc/base.py
