+++ services/assistant/app/adapters/output/grpc/xray.py
@@ -3,7 +3,7 @@ from app.contracts.clients import XrayClient
 from app.adapters.output.grpc.gen.xray_pb2_grpc import XrayStub
 from app.adapters.output.grpc.gen.xray_pb2 import RestartResponse, Null
 
-from app.adapters.output.grpc.grpc import GRPCClient
+from app.adapters.output.grpc.base import GRPCClient
 
 
 class GRPCXrayClient(GRPCClient, XrayClient):
